% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osler_install.R
\name{osler_install}
\alias{osler_install}
\alias{oslerLite}
\title{OSLER Installer}
\usage{
osler_install(repo, release = c("stable", "current"),
  upgrade_dependencies = FALSE, ...)

oslerLite(...)
}
\arguments{
\item{repo}{Package name in OSLER}

\item{release}{Stable or development version}

\item{upgrade_dependencies}{Should dependencies be updated?
passed to \code{\link[devtools]{install}}}

\item{...}{additional arguments passed to
\code{\link[devtools]{install_github}}}
}
\value{
Result from \code{\link[devtools]{install_github}}
}
\description{
Install function for OSLER packages
}
