% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osler_package_table.R
\name{osler_packages}
\alias{osler_packages}
\title{OSLER Packages}
\usage{
osler_packages(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{osler_package_table}}}
}
\value{
\code{vector} of packages available on OSLER
}
\description{
Returns the vector of OSLER packages
}
\examples{
osler_packages()
}
